Implement CLI
Using Boost Program Options library. There's still a need for refactoring and
error handling though.

+++ ex-2-images/ImageDataProviderInterface.cpp
@@ -9,5 +9,5 @@ float ImageDataProviderInterface::mapTo0To1Range(int value, int maxColorValue) c
     if (value < 0 || maxColorValue < value) {
         throw IncorrectColorValueException(maxColorValue);
     }
-    return value / maxColorValue;
+    return (double) value / maxColorValue;
 }
